Refactor node generation fetching with SWR
Replaced custom fetching implementation with SWR for better data management:
- Moved node generation fetching logic from context to hook
- Added SWR for more efficient data fetching and caching
- Exposed setGenerations in context for better state management
- Updated types to support the new implementation
